task - 2 - fixed balance type

+++ task_2/BankAccount.java
@@ -1,21 +1,21 @@
 package task_2;
 
 public class BankAccount {
-    private int balance;
+    private double balance;
 
-    public BankAccount(int _balance) {
+    public BankAccount(double _balance) {
         balance = _balance;
     }
 
-    public void deposit(int value) {
+    public void deposit(double value) {
         balance += value;
     }
 
-    public void withdraw(int value) {
+    public void withdraw(double value) {
         balance -= value;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 }
